refactor: рефакторинг страницы Places по замечаниям
