+++ src/pages/Places/index.js
@@ -2,13 +2,13 @@ import BasePage from '../../layout/BasePage';
 import CardPlace from '../../components/Cards/CardPlace/CardPlace';
 import PlacesRecommend from '../../components/PlacesRecommend/PlacesRecommend';
 import AnimatedPageContainer from '../../components/AnimatedPageContainer/AnimatedPageContainer';
-import {
-  TitleH1,
-  NoDataNotificationBox,
-  Loader,
-  Paginate,
-} from '../../components/utils/index';
 import TagsList from '../../components/TagsList/TagsList';
+import {
+  Loader,
+  NoDataNotificationBox,
+  Paginate,
+  TitleH1,
+} from '../../components/utils';
 
 export {
   BasePage,
